Working on Real Project with Python (cars Dataset) (A Part of Big Data Analyst )
# Working on Real Project with Python (cars Dataset)
(A Part of Big Data Analyst )
Cars Dataset
Here, the data of Differenr is given with their specifications.¶
this data is availables as CSV file, we are going to analysze this data set using the pandas DataFrame.

1. Instruction  (For data Cleaning)
Find  all Null Values in the  dataset. If there is any null values in any columns ,then fill the mean of that columns.
2.    Questions ( Based on value counts)
Check what are different type of make are there in our dataset. And what is the count 
(Occurrence ) of each Make I the data ?
3.  instrusction (Filtering )
	show all the record ,where origin is Asia or Europe
4.  instruction (Removing  unwanted record)
	Remove all the record (rows) where weight is above 4000.
5. instructions (Applying funtion on a column )
      	  increase all the values of " MPG_City" column by 3